change regs to reg_t

+++ include_winix/winix/slab.h
@@ -19,7 +19,7 @@
 void* kmalloc(unsigned int size);
 void kfree(void *ptr);
 
-void* _kmalloc(unsigned int size, void *ra);
+void* _kmalloc(size_t size, void *ra);
 void _kfree(void *ptr, void *ra);
 void* krealloc(void *p,unsigned int size);
 void *kcalloc(unsigned int number , unsigned int size);
